Fix bug where we deleted the wrong node, misc file cleanups
