--- tests/integration/cartography/intel/automated_schema_tester.py
@@ -1,10 +1,0 @@
-def build_tester_queries(node_schema: CartographyNodeSchema) -> List[str]:
-    return []
-
-
-def test_all_nodes(neo4j_session):
-    # Get all schemas from the models folder
-
-    queries = build_tester_queries()
-
-    #for query in queries:
